setup: update

+++ setup/linking.fish
@@ -15,6 +15,7 @@ for config in alacritty aria2 \
     mgba mpv \
     newsboat niri npm nvim \
     paru python \
+    qBittorrent \
     ruff \
     sqlite3 streamrip systemd \
     termux tmux \
@@ -22,6 +23,7 @@ for config in alacritty aria2 \
     xdg-desktop-portal xdg-desktop-portal-termfilechooser \
     yay yazi yt-dlp \
     zathura zellij
+
     # skip non-existing dirs
     [ -e ~/repos/dotfiles/$config ] || continue
     # unlink/relocate old directories in ~/.config
@@ -33,16 +35,20 @@ end
 printf "done\n"
 
 printf "LINKING CONFIG FILES...\n"
-ln -fs ~/repos/dotfiles/other/clang-format ~/.clang-format
+ln -fs ~/repos/dotfiles/other/clang-format ~/repos/.clang-format
+ln -fs ~/repos/dotfiles/other/clang-format ~/.local/.clang-format
 ln -fs ~/repos/dotfiles/other/curlrc ~/.config/.curlrc
-ln -fs ~/repos/dotfiles/other/tidal-dl.json ~/.config/.tidal-dl.json
 ln -fs ~/repos/dotfiles/other/starship.toml ~/.config/starship.toml
 ln -fs ~/repos/dotfiles/other/stylua.toml ~/.config/stylua.toml
 ln -fs ~/repos/dotfiles/other/taplo.toml ~/.config/taplo.toml
-ln -fs ~/repos/dotfiles/other/ssh_config ~/.ssh/config
+# ln -fs ~/repos/dotfiles/other/ssh_config ~/.ssh/config # do this manually
 
-if [ -e /etc/pacman.conf ]
-    sudo ln -fs ~/repos/dotfiles/other/pacman.arch.conf /etc/pacman.conf
+[ -e /etc/pacman.conf ] && sudo ln -fs ~/repos/dotfiles/other/pacman.arch.conf /etc/pacman.conf
+
+if [ (uname -o) = Android ]
+    ln -fs ~/repos/dotfiles/other/pacman.termux.conf $PREFIX/etc/pacman.conf
+    ln -fs ~/repos/dotfiles/termux/colors.properties ~/.termux/colors.properties # https://github.com/termux/termux-app/blob/master/termux-shared/src/main/java/com/termux/shared/termux/TermuxConstants.java#L657
+    ln -fs ~/repos/dotfiles/termux/termux.properties ~/.termux/termux.properties
 end
 
 if [ (uname -o) = Linux ]
@@ -50,12 +56,6 @@ if [ (uname -o) = Linux ]
 end
 
 if [ (uname -o) = Android ]
-    # configs
-    ln -fs ~/repos/dotfiles/other/pacman.termux.conf $PREFIX/etc/pacman.conf
-    ln -fs ~/repos/dotfiles/termux/colors.properties ~/.termux/colors.properties # https://github.com/termux/termux-app/blob/master/termux-shared/src/main/java/com/termux/shared/termux/TermuxConstants.java#L657
-    ln -fs ~/repos/dotfiles/termux/termux.properties ~/.termux/termux.properties
-
-    # scripts
     ln -fs ~/repos/dotfiles/scripts/bin/termux-url-opener ~/bin/termux-url-opener
     ln -fs ~/repos/dotfiles/scripts/bin/termux-file-editor ~/bin/termux-file-editor
     ln -fs ~/repos/dotfiles/scripts/bin/rish ~/.local/bin/rish
